chtml/html/testdata/go/template.dat

#data
<body><template><yt-icon-button></yt-icon-button><form><paper-input></paper-input></form><style></style></template>
#errors
#document
| <body>
|   <template>
|     content
|       <yt-icon-button>
|       <form>
|         <paper-input>
|       <style>

#data
<template><tBody><isindex/action=0>
#errors
#document
| <template>
|   content
|     <tbody>
|     <isindex>
|       action="0"

#data
<math><template><mo><template>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math template>
|         <math mo>
|           <template>
|             content

#data
<svg><template><desc><t><svg></template>
#errors
#document
| <svg svg>
|   <svg template>
|     <svg desc>
|       <t>
|         <svg svg>

#data
<math><template><mn><b></template>
#errors
#document
| <math math>
|   <math template>
|     <math mn>
|       <b>
